+++ changelog.md
@@ -1,6 +1,10 @@
 Change Log
 ==========
 
+HEAD
+----
+
+ * Added `getColumn(String).as(Class<T>)` method returning deserialization `Function<Cursor, T>` which uses custom type adapters registered in MicroOrm instance.
 
 Version 0.5.0 *(2014-09-19)*
 ----------------------------